Add support for AnimeBytes.tv scrape.php
